refactor(040): migra packages/core/src/markdown para TypeScript
