+++ packages/core/src/markdown/index.ts
@@ -1,2 +1,2 @@
 // @dosiq/core/markdown — parse leve de markdown do chatbot (paridade web↔mobile, spec 015).
-export { parseMessageMarkdown, parseInlineSegments } from './parseMessageMarkdown.js'
+export { parseMessageMarkdown, parseInlineSegments } from './parseMessageMarkdown'
